@@ -15,6 +15,7 @@
     "lint": "prettier --plugin prettier-plugin-svelte . --check . && eslint .",
     "format": "prettier --plugin prettier-plugin-svelte . --write .",
     "docgen-props": "NODE_OPTIONS='--loader ts-node/esm/transpile-only --no-warnings' node scripts/docgenProps.ts",
+    "colorgen": "NODE_OPTIONS='--loader ts-node/esm/transpile-only --no-warnings' node ./src/dev/colorgenPlugin.ts",
     "docgen-snippets": "NODE_OPTIONS='--loader ts-node/esm/transpile-only --no-warnings' node scripts/docgenSnippets.ts",
     "write-version": "NODE_OPTIONS='--loader ts-node/esm/transpile-only --no-warnings' node scripts/writeVersion.ts"
   },
@@ -34,30 +35,32 @@
     "svelte": "^4.0.0"
   },
   "devDependencies": {
-    "@fontsource/roboto": "^5.0.8",
-    "@sveltejs/adapter-auto": "^2.1.0",
-    "@sveltejs/kit": "^1.25.2",
-    "@sveltejs/package": "^2.2.2",
-    "@types/node": "^20.8.6",
+    "@fontsource/roboto": "^5.0.13",
+    "@material/material-color-utilities": "^0.2.7",
+    "@sveltejs/adapter-auto": "^2.1.1",
+    "@sveltejs/kit": "^1.30.4",
+    "@sveltejs/package": "^2.3.1",
+    "@types/node": "^20.12.8",
     "@types/prettier": "^2.7.3",
-    "@types/prismjs": "^1.26.1",
-    "@typescript-eslint/eslint-plugin": "^6.8.0",
-    "@typescript-eslint/parser": "^6.8.0",
-    "eslint": "^8.51.0",
-    "eslint-config-prettier": "^9.0.0",
-    "eslint-plugin-svelte": "^2.34.0",
+    "@types/prismjs": "^1.26.3",
+    "@typescript-eslint/eslint-plugin": "^6.21.0",
+    "@typescript-eslint/parser": "^6.21.0",
+    "eslint": "^8.57.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-svelte": "^2.38.0",
     "eslint-plugin-unicorn": "^48.0.1",
-    "material-symbols": "^0.13.1",
-    "prettier": "^3.0.3",
-    "prettier-plugin-svelte": "^3.0.3",
-    "publint": "^0.2.4",
-    "sass": "^1.69.4",
-    "svelte": "^4.2.1",
-    "svelte-check": "^3.5.2",
-    "ts-node": "^10.9.1",
+    "magic-string": "^0.30.10",
+    "material-symbols": "^0.13.2",
+    "prettier": "^3.2.5",
+    "prettier-plugin-svelte": "^3.2.3",
+    "publint": "^0.2.7",
+    "sass": "^1.76.0",
+    "svelte": "5.0.0-next.123",
+    "svelte-check": "^3.7.1",
+    "ts-node": "^10.9.2",
     "tslib": "^2.6.2",
-    "typescript": "^5.2.2",
-    "vite": "^4.4.11",
+    "typescript": "^5.4.5",
+    "vite": "^4.5.3",
     "vitest": "^0.34.6"
   },
   "svelte": "./dist/index.js",
@@ -65,7 +68,7 @@
   "type": "module",
   "dependencies": {
     "highlight.js": "^11.9.0",
-    "material-dynamic-colors": "^1.0.5",
-    "svelte-highlight": "^7.4.1"
+    "material-dynamic-colors": "^1.1.0",
+    "svelte-highlight": "^7.6.0"
   }
 }
