@@ -1,69 +1,42 @@
+<svelte:options runes={true} />
+
 <script lang="ts">
-  import { useSize } from "$lib/composables";
+  import { useSize } from "$lib/composables/useSize";
   import type { QAvatarProps } from "./props";
 
-  export let shape: QAvatarProps["shape"] = "circle",
-    size: QAvatarProps["size"] = "md",
-    src: QAvatarProps["src"] = undefined,
-    video: QAvatarProps["video"] = false,
-    userClasses: QAvatarProps["userClasses"] = "";
-  export { userClasses as class };
+  let {
+    alt,
+    shape = "circle",
+    size = "md",
+    src,
+    video = false,
+    children,
+    videoAccessibility,
+    ...props
+  }: QAvatarProps = $props();
 
-  $: sizeObj = useSize(size);
+  const qSize = $derived(useSize(size, "q-avatar"));
+  const qShape = $derived(`q-avatar--${shape}`);
+
+  Q.classes("q-avatar", {
+    classes: [qShape, qSize.class, props.class],
+  });
 </script>
 
-{#if video === true}
-  <!-- svelte-ignore a11y-media-has-caption -->
-  <video
-    class="q-avatar {sizeObj.class ? `q-avatar--${sizeObj.class}` : ''} {userClasses}"
-    class:q-avatar--circle={shape === "circle"}
-    class:q-avatar--round={shape === "rounded"}
-    class:q-avatar--top-round={shape?.includes("top")}
-    class:q-avatar--bottom-round={shape?.includes("bottom")}
-    class:q-avatar--left-round={shape?.includes("left")}
-    class:q-avatar--right-round={shape?.includes("right")}
-    style:width={sizeObj.style}
-    style:height={sizeObj.style}
-    autoplay
-    loop
-    playsinline
-    {...$$restProps}
-    on:click
-  >
-    <source {src} type="video/mp4" />
-  </video>
-{:else if src !== undefined}
-  <!-- svelte-ignore a11y-missing-attribute -->
-  <!-- svelte-ignore a11y-no-noninteractive-element-interactions -->
-  <img
-    class="q-avatar {sizeObj.class ? `q-avatar--${sizeObj.class}` : ''} {userClasses}"
-    class:q-avatar--circle={shape === "circle"}
-    class:q-avatar--round={shape === "rounded"}
-    class:q-avatar--top-round={shape?.includes("top")}
-    class:q-avatar--bottom-round={shape?.includes("bottom")}
-    class:q-avatar--left-round={shape?.includes("left")}
-    class:q-avatar--right-round={shape?.includes("right")}
-    style:width={sizeObj.style}
-    style:height={sizeObj.style}
-    {src}
-    {...$$restProps}
-    on:click
-  />
-{:else}
-  <!-- svelte-ignore a11y-no-static-element-interactions -->
-  <div
-    class="q-avatar {sizeObj.class ? `q-avatar--${sizeObj.class}` : ''} {userClasses}"
-    class:q-avatar--circle={shape === "circle"}
-    class:q-avatar--round={shape === "rounded"}
-    class:q-avatar--top-round={shape?.includes("top")}
-    class:q-avatar--bottom-round={shape?.includes("bottom")}
-    class:q-avatar--left-round={shape?.includes("left")}
-    class:q-avatar--right-round={shape?.includes("right")}
-    style:width={sizeObj.style}
-    style:height={sizeObj.style}
-    {...$$restProps}
-    on:click
-  >
-    <slot />
-  </div>
-{/if}
+<div {...props} class="q-avatar" {...Q.classes} style:--size={qSize.style}>
+  {#if video}
+    <video>
+      <source {src} type="video/mp4" />
+      <track kind="captions" />
+      {@render videoAccessibility?.()}
+    </video>
+  {:else if src !== undefined}
+    <img {src} {alt} />
+  {:else}
+    {@render children?.()}
+  {/if}
+</div>
+
+<style lang="scss">
+  @import "./QAvatar.scss";
+</style>
