@@ -1,22 +1,34 @@
+<svelte:options runes={true} />
+
 <script lang="ts">
-  import { onMount, setContext } from "svelte";
+  import { setContext, untrack } from "svelte";
   import type { QBreadcrumbsProps } from "./props";
 
-  export let separator: QBreadcrumbsProps["separator"] = "/",
-    gutter: QBreadcrumbsProps["gutter"] = "sm",
-    activeColor: QBreadcrumbsProps["activeColor"] = "primary",
-    separatorColor: QBreadcrumbsProps["separatorColor"] = "outline",
-    userClasses: QBreadcrumbsProps["userClasses"] = "";
-  export { userClasses as class };
+  let {
+    activeColor = "primary",
+    gutter = "sm",
+    separator = "/",
+    separatorColor = "outline",
+    children,
+    ...props
+  }: QBreadcrumbsProps = $props();
 
-  let breadcrumbElement: HTMLDivElement;
+  let breadrumbElement: HTMLDivElement;
 
-  onMount(() => breadcrumbElement.firstChild?.remove());
+  $effect(() => {
+    untrack(() => breadrumbElement.firstChild?.remove());
+  });
 
   setContext("activeColor", activeColor);
   setContext("separator", { type: separator, color: separatorColor, gutter });
+
+  Q.classes("q-breadcrumbs", { classes: [props.class] });
 </script>
 
-<div class="q-breadcrumbs {userClasses}" bind:this={breadcrumbElement}>
-  <slot />
+<div bind:this={breadrumbElement} {...props} class="q-breadcrumbs" {...Q.classes}>
+  {@render children?.()}
 </div>
+
+<style lang="scss">
+  @import "./QBreadcrumbs.scss";
+</style>
