@@ -1,34 +1,41 @@
 <script lang="ts">
-  import { onMount, setContext } from "svelte";
+  import { QContext } from "$lib/classes/QContext.svelte";
   import type { QListProps } from "./props";
 
-  export let bordered: QListProps["bordered"] = false,
-    roundedBorders: QListProps["roundedBorders"] = false,
-    dense: QListProps["dense"] = false,
-    separator: QListProps["separator"] = false,
-    separatorOptions: QListProps["separatorOptions"] = {},
-    padding: QListProps["padding"] = false,
-    tag: QListProps["tag"] = "div",
-    userClasses: QListProps["userClasses"] = "";
-  export { userClasses as class };
+  let {
+    bordered = false,
+    roundedBorders = false,
+    dense = false,
+    separator = false,
+    separatorOptions = {},
+    padding = false,
+    tag = "div",
+    children,
+    ...props
+  }: QListProps = $props();
 
-  let listElement: HTMLElement;
+  let listEl: HTMLElement;
 
-  onMount(() => listElement.querySelector(".q-separator__wrapper:first-child")?.remove());
+  $effect(() => {
+    listEl.querySelector(".q-separator__wrapper:first-child")?.remove();
+  });
 
-  $: setContext("separator", separator === true ? separatorOptions : undefined);
+  new QContext(separator ? separatorOptions : undefined).set("separator");
+
+  Q.classes("q-list", {
+    bemClasses: {
+      bordered,
+      dense,
+      rounded: roundedBorders,
+    },
+    classes: [padding && "q-py-sm", "flex", "flex-center", props.class],
+  });
 </script>
 
-<svelte:element
-  this={tag}
-  class="q-list {userClasses}"
-  class:q-list--bordered={bordered}
-  class:q-list--rounded={roundedBorders}
-  class:q-list--dense={dense}
-  class:q-py-sm={padding}
-  {...$$restProps}
-  bind:this={listElement}
-  on:scroll
->
-  <slot />
+<svelte:element this={tag} bind:this={listEl} {...props} class="q-list" {...Q.classes}>
+  {@render children?.()}
 </svelte:element>
+
+<style lang="scss">
+  @import "./QList.scss";
+</style>
