@@ -1,5 +1,5 @@
 <script lang="ts">
-  import { QCodeBlock, QDialog } from "$lib";
+  import { QCodeBlock, QDialog, QBtn } from "$lib";
 
   export let title: string,
     snippets: Record<string, string> | undefined = undefined;
@@ -11,14 +11,8 @@
   <div class="flex justify-between q-mb-md">
     <h5>{title}</h5>
     {#if snippets}
-      <QDialog
-        class="snippet-dialog"
-        bind:value={dialog}
-        button
-        buttonProps={{ outline: true, round: true, icon: "code" }}
-        on:buttonClick={() => (dialog = true)}
-        modal
-      >
+      <QBtn icon="code" design="outlined" round onclick={() => (dialog = true)} />
+      <QDialog class="snippet-dialog" bind:value={dialog} modal>
         <QCodeBlock code={snippets[title]} language="svelte" {title} copiable />
       </QDialog>
     {/if}
