@@ -1,26 +1,30 @@
+<svelte:options runes={true} />
+
 <script lang="ts">
   import type { QToolbarProps } from "./props";
 
-  export let inset: QToolbarProps["inset"] = false,
-    border: QToolbarProps["border"] = false,
-    elevate: QToolbarProps["elevate"] = false,
-    height: QToolbarProps["height"] = "64px",
-    userClasses: QToolbarProps["userClasses"] = "";
-  export { userClasses as class };
+  let {
+    inset = false,
+    border = false,
+    elevate = false,
+    height = "64px",
+    children,
+    ...props
+  }: QToolbarProps = $props();
 
-  $: dynamicHeight = !userClasses?.includes("q-header") ? height : undefined;
+  const dynamicHeight = $derived(!props.class?.includes("q-header") ? height : undefined);
 </script>
 
 <header
-  class="q-toolbar {userClasses}"
+  {...props}
+  class="q-toolbar {props.class}"
   class:q-toolbar--inset={inset}
   class:q-toolbar--elevated={elevate}
   class:q-toolbar--bordered={border}
   role="toolbar"
   style:height={dynamicHeight}
-  {...$$restProps}
 >
   <nav>
-    <slot />
+    {@render children?.()}
   </nav>
 </header>
