@@ -66,7 +66,7 @@
 <QDocs componentDocs={QLayoutDocs}>
   <QLayout slot="display" view="lhh lpr lfr" headerHeight="50px" footerHeight="50px">
     <QHeader slot="header" elevate>
-      <QBtn icon="menu" flat on:click={displayLeftDrawerElement.toggle} />
+      <QBtn icon="menu" design="flat" onclick={displayLeftDrawerElement.toggle} />
       <QToolbarTitle>Header</QToolbarTitle>
     </QHeader>
     <QDrawer
@@ -119,7 +119,7 @@
         <QLayout {view} {style}>
           <QHeader style={header ? undefined : "display: none;"} slot="header" elevate>
             {#if leftDrawer}
-              <QBtn icon="menu" flat on:click={leftDrawerElement.toggle} />
+              <QBtn icon="menu" design="flat" onclick={leftDrawerElement.toggle} />
             {/if}
             <div class="flex column">
               <QRadio bind:selected={viewArr[0][0]} value="h" label="h" />
@@ -144,7 +144,7 @@
               <QRadio style="margin: 0" bind:selected={viewArr[0][2]} value="r" label="r" />
             </div>
             {#if rightDrawer}
-              <QBtn icon="menu" flat on:click={rightDrawerElement.toggle} />
+              <QBtn icon="menu" design="flat" onclick={rightDrawerElement.toggle} />
             {/if}
           </QHeader>
 
@@ -252,7 +252,7 @@
 
           <QFooter style={footer ? undefined : "display: none;"} slot="footer" elevate>
             {#if leftDrawer}
-              <QBtn icon="menu" flat on:click={leftDrawerElement.toggle} />
+              <QBtn icon="menu" design="flat" onclick={leftDrawerElement.toggle} />
             {/if}
             <div class="flex column">
               <QRadio bind:selected={viewArr[2][0]} value="f" label="f" />
@@ -277,7 +277,7 @@
               <QRadio style="margin: 0" bind:selected={viewArr[2][2]} value="r" label="r" />
             </div>
             {#if rightDrawer}
-              <QBtn icon="menu" flat on:click={rightDrawerElement.toggle} />
+              <QBtn icon="menu" design="flat" onclick={rightDrawerElement.toggle} />
             {/if}
           </QFooter>
 
