@@ -1,11 +1,9 @@
-import { NativePropsDefaults } from "$lib/utils";
-import { UseRouterLinkPropsDefaults } from "$lib/composables/useRouterLink";
-import type { UseRouterLinkProps } from "$lib/composables/useRouterLink";
+import type { RouterProps } from "$lib/utils/router";
+import type { Snippet } from "svelte";
 import type { QSeparatorProps } from "../separator/props";
 import type { HTMLAnchorAttributes, HTMLAttributes } from "svelte/elements";
-import type { NativeProps } from "$lib/utils";
 
-export interface QListProps extends NativeProps, HTMLAttributes<HTMLElement> {
+export interface QListProps extends HTMLAttributes<HTMLElement> {
   bordered?: boolean;
   roundedBorders?: boolean;
   dense?: boolean;
@@ -22,18 +20,7 @@ export interface QListProps extends NativeProps, HTMLAttributes<HTMLElement> {
   tag?: string;
 }
 
-export const QListPropsDefaults: QListProps = {
-  bordered: false,
-  roundedBorders: false,
-  dense: false,
-  separator: false,
-  separatorOptions: {},
-  padding: false,
-  tag: "div",
-  ...NativePropsDefaults,
-};
-
-export interface QItemProps extends UseRouterLinkProps, NativeProps, HTMLAttributes<HTMLElement> {
+export interface QItemProps extends RouterProps, HTMLAttributes<HTMLElement> {
   tag?: string;
   active?: boolean;
   clickable?: boolean;
@@ -42,16 +29,6 @@ export interface QItemProps extends UseRouterLinkProps, NativeProps, HTMLAttribu
   tabindex?: HTMLAttributes<HTMLElement>["tabindex"];
   target?: HTMLAnchorAttributes["target"];
 }
-
-export const QItemPropsDefaults: QItemProps = {
-  tag: "div",
-  active: false,
-  clickable: false,
-  dense: false,
-  tabindex: 0,
-  ...UseRouterLinkPropsDefaults,
-  ...NativePropsDefaults,
-};
 
 export type QItemSectionTypes =
   | "thumbnail"
@@ -63,11 +40,10 @@ export type QItemSectionTypes =
   | "trailingText"
   | "content";
 
-export interface QItemSectionProps extends NativeProps, HTMLAttributes<HTMLDivElement> {
+export interface QItemSectionProps extends HTMLAttributes<HTMLDivElement> {
   type?: QItemSectionTypes;
+  headline?: Snippet;
+  line1?: Snippet;
+  line2?: Snippet;
+  line3?: Snippet;
 }
-
-export const QItemSectionPropsDefaults: QItemSectionProps = {
-  type: "content",
-  ...NativePropsDefaults,
-};
