@@ -4,7 +4,7 @@
 
 <div class="row q-gutter-lg" style="min-height: 400px; margin: 1rem">
   <QCard class="col-sm-12 col-md-6 flex primary-container">
-    <h1 class="large" slot="title">The $Quaff object</h1>
+    <h1 class="large">The $Quaff object</h1>
   </QCard>
   <QCard class="col-sm-12 col-md-6 q-mt-none secondary-container flex">
     <QCardSection class="flex">
