@@ -1,11 +1,12 @@
 import adapter from "@sveltejs/adapter-auto";
 import { vitePreprocess } from "@sveltejs/kit/vite";
+import { preprocessClasses } from "./plugins/class-preprocessor/index.js";
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
   // Consult https://kit.svelte.dev/docs/integrations#preprocessors
   // for more information about preprocessors
-  preprocess: vitePreprocess(),
+  preprocess: [vitePreprocess(), preprocessClasses("Q")],
 
   experimental: {
     useVitePreprocess: true,
